+++ Extensions/Minded.Extensions.Caching.Abstractions/Decorator/IGenerateCacheKey.cs
@@ -1,0 +1,15 @@
+﻿namespace Minded.Extensions.Caching.Decorator
+{
+    /// <summary>
+    /// This interface can be used to generate a custom cache key
+    /// The key must be unique.
+    /// </summary>
+    public interface IGenerateCacheKey
+    {
+        /// <summary>
+        /// Returns the unique cache key
+        /// </summary>
+        /// <returns>String cache key</returns>
+        string GetCacheKey();
+    }
+}