+++ Extensions/Minded.Extensions.Caching.Abstractions/Decorator/NullGlobalCacheKeyPrefixProvider.cs
@@ -1,0 +1,9 @@
+﻿using Minded.Extensions.Caching.Decorator;
+
+namespace Minded.Extensions.Caching.Abstractions.Decorator
+{
+    public class NullGlobalCacheKeyPrefixProvider : IGlobalCacheKeyPrefixProvider
+    {
+        public string GetGlobalCacheKeyPrefix() => "";
+    }
+}